# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"SupplierDto": {
				"description": "DTO pro zobrazení dodavatele\nObsahuje všechny properties dodavatele včetně ID a timestamps",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"contactName": {
						"type": "string"
					},
					"contactEmail": {
						"type": "string"
					},
					"contactPhone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"isActive",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateSupplierDto": {
				"description": "Data Transfer Object (DTO) pro vytvoření nového dodavatele",
				"properties": {
					"name": {
						"type": "string"
					},
					"contactName": {
						"type": "string"
					},
					"contactEmail": {
						"type": "string"
					},
					"contactPhone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateSupplierDto": {
				"description": "DTO pro aktualizaci existujícího dodavatele\nVšechny properties jsou volitelné",
				"properties": {
					"name": {
						"type": "string"
					},
					"contactName": {
						"type": "string"
					},
					"contactEmail": {
						"type": "string"
					},
					"contactPhone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ProductImageDto": {
				"description": "DTO pro zobrazení obrázku produktu\nObsahuje všechny properties obrázku produktu včetně ID a timestamps",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"productId": {
						"type": "number",
						"format": "double"
					},
					"url": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"altText": {
						"type": "string"
					},
					"sortOrder": {
						"type": "number",
						"format": "double"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"imageType": {
						"type": "string"
					},
					"isMain": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"productId",
					"url",
					"sortOrder",
					"isMain",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DbErrorResponse": {
				"description": "Rozhraní pro odpověď s informací o chybě databáze",
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"recovered": {
						"type": "boolean"
					}
				},
				"required": [
					"success",
					"message",
					"error",
					"recovered"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "express-oracle-api",
		"version": "1.0.0",
		"description": "REST API s Express.js, TypeORM, Zod a Oracle",
		"contact": {}
	},
	"paths": {
...
		"/suppliers": {
			"get": {
				"operationId": "GetSuppliers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalPages": {
											"type": "number",
											"format": "double"
										},
										"limit": {
											"type": "number",
											"format": "double"
										},
										"page": {
											"type": "number",
											"format": "double"
										},
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {
											"items": {
												"$ref": "#/components/schemas/SupplierDto"
											},
											"type": "array"
										}
									},
									"required": [
										"totalPages",
										"limit",
										"page",
										"total",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Získá seznam všech dodavatelů s podporou filtrování, řazení a stránkování",
				"tags": [
					"Suppliers"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateSupplier",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SupplierDto"
								}
							}
						}
					}
				},
				"description": "Vytvoří nového dodavatele",
				"tags": [
					"Suppliers"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Data pro vytvoření dodavatele",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSupplierDto",
								"description": "Data pro vytvoření dodavatele"
							}
						}
					}
				}
			}
		},
		"/suppliers/{id}": {
			"get": {
				"operationId": "GetSupplier",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SupplierDto"
								}
							}
						}
					}
				},
				"description": "Získá dodavatele podle ID",
				"tags": [
					"Suppliers"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID dodavatele",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateSupplier",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SupplierDto"
								}
							}
						}
					}
				},
				"description": "Aktualizuje existujícího dodavatele",
				"tags": [
					"Suppliers"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID dodavatele",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Data pro aktualizaci dodavatele",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSupplierDto",
								"description": "Data pro aktualizaci dodavatele"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteSupplier",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Odstraní dodavatele",
				"tags": [
					"Suppliers"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID dodavatele",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/suppliers/country/{country}": {
			"get": {
				"operationId": "GetSuppliersByCountry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SupplierDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Získá dodavatele podle země",
				"tags": [
					"Suppliers"
				],
				"security": [],
				"parameters": [
					{
						"description": "Země dodavatele",
						"in": "path",
						"name": "country",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/product-images/product/{productId}": {
			"get": {
				"operationId": "GetProductImagesByProductId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductImageDto"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Chyba při načítání obrázků produktu",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					}
				},
				"description": "Získá všechny obrázky pro konkrétní produkt",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID produktu",
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/product-images/{id}/main/product/{productId}": {
			"put": {
				"operationId": "SetMainImage",
				"responses": {
					"200": {
						"description": "Obrázek byl nastaven jako hlavní",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Obrázek nebyl nalezen",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Chyba při nastavování hlavního obrázku",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					}
				},
				"description": "Nastaví obrázek jako hlavní pro konkrétní produkt",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": [
					{
						"description": "ID obrázku",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "ID produktu",
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/product-images/seed": {
			"post": {
				"operationId": "SeedTestData",
				"responses": {
					"201": {
						"description": "Testovací data úspěšně seedována",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"message",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Chyba při seedování testovacích dat",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DbErrorResponse"
								}
							}
						}
					}
				},
				"description": "Seedování testovacích dat obrázků produktů",
				"tags": [
					"Obrázky produktů"
				],
				"security": [],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}